#132: Restructure Iterator trait to split out instrumentation logic and reduce code duplication
