+++ src/iterator/component_iterator.rs
@@ -5,11 +5,12 @@ use crate::ir::id::{FunctionID, GlobalID, LocalID, ModuleID};
 use crate::ir::module::module_functions::FuncKind;
 use crate::ir::module::module_globals::Global;
 use crate::ir::types::{DataType, Instrument, InstrumentType, InstrumentationMode, Location};
-use crate::iterator::iterator_trait::Iterator;
+use crate::iterator::iterator_trait::{Instrumenter, Iterator, set_instrument_type_for_local_func_at};
 use crate::opcode::{Inject, MacroOpcode, Opcode};
 use crate::subiterator::component_subiterator::ComponentSubIterator;
 use crate::ModuleBuilder;
 use std::collections::HashMap;
+use std::iter::Iterator as StdIter;
 use wasmparser::Operator;
 
 /// Iterator for a Component.
@@ -120,7 +121,7 @@ impl<'a, 'b> Inject<'b> for ComponentIterator<'a, 'b> {
     /// use orca::iterator::component_iterator::ComponentIterator;
     /// use wasmparser::Operator;
     /// use orca::ir::types::{Location};
-    /// use orca::iterator::iterator_trait::Iterator;
+    /// use orca::iterator::iterator_trait::{Instrumenter, Iterator};
     /// use orca::opcode::Opcode;
     ///
     /// let file = "path_to_file";
@@ -175,44 +176,7 @@ impl<'a, 'b> Inject<'b> for ComponentIterator<'a, 'b> {
 }
 impl<'a, 'b> Opcode<'b> for ComponentIterator<'a, 'b> {}
 impl<'a, 'b> MacroOpcode<'b> for ComponentIterator<'a, 'b> {}
-
-// Note: Marked Trait as the same lifetime as component
-impl<'a, 'b> Iterator<'b> for ComponentIterator<'a, 'b> {
-    /// Marks the current location as InstrumentBefore
-    fn before(&mut self) -> &mut Self {
-        self.set_instrument_type(InstrumentationMode::Before);
-        self
-    }
-
-    /// Marks the current location as InstrumentAfter
-    fn after(&mut self) -> &mut Self {
-        self.set_instrument_type(InstrumentationMode::After);
-        self
-    }
-
-    /// Marks the current location as InstrumentAlternate
-    fn alternate(&mut self) -> &mut Self {
-        self.set_instrument_type(InstrumentationMode::Alternate);
-        self
-    }
-
-    /// Resets the Component Iterator
-    fn reset(&mut self) {
-        self.comp_iterator.reset();
-    }
-
-    /// Goes to the next instruction
-    fn next(&mut self) -> Option<&Operator> {
-        match self.comp_iterator.next() {
-            false => None,
-            true => self.curr_op(),
-        }
-    }
-
-    /// Returns the current location
-    fn curr_loc(&self) -> Location {
-        self.comp_iterator.curr_loc()
-    }
+impl <'a, 'b> Instrumenter<'b> for ComponentIterator<'a, 'b> {
 
     /// Returns the Instrumentation at the current Location
     fn curr_instrument_type(&self) -> InstrumentType {
@@ -237,50 +201,6 @@ impl<'a, 'b> Iterator<'b> for ComponentIterator<'a, 'b> {
         }
     }
 
-    /// Returns the instruction at the current location
-    fn curr_op(&self) -> Option<&Operator> {
-        if self.comp_iterator.end() {
-            None
-        } else if let Location::Component {
-            mod_idx,
-            func_idx,
-            instr_idx,
-        } = self.comp_iterator.curr_loc()
-        {
-            match &self.comp.modules[mod_idx]
-                .functions
-                .get(func_idx as FunctionID)
-                .kind
-            {
-                FuncKind::Import(_) => panic!("Can't inject into an imported function!"),
-                FuncKind::Local(l) => Some(&l.body.instructions[instr_idx].0),
-            }
-        } else {
-            panic!("Should have gotten Component Location and not Module Location!")
-        }
-    }
-
-    /// Gets the injected instruction at the current location by index
-    fn get_injected_val(&self, idx: usize) -> &Operator {
-        if let Location::Component {
-            mod_idx,
-            func_idx,
-            instr_idx,
-        } = self.comp_iterator.curr_loc()
-        {
-            match &self.comp.modules[mod_idx]
-                .functions
-                .get(func_idx as FunctionID)
-                .kind
-            {
-                FuncKind::Import(_) => panic!("Can't inject into an imported function!"),
-                FuncKind::Local(l) => l.body.instructions[instr_idx].1.get_instr(idx),
-            }
-        } else {
-            panic!("Should have gotten Component Location and not Module Location!")
-        }
-    }
-
     /// Sets the type of Instrumentation Mode of the current location
     fn set_instrument_type(&mut self, mode: InstrumentationMode) {
         if let Location::Component {
@@ -290,6 +210,28 @@ impl<'a, 'b> Iterator<'b> for ComponentIterator<'a, 'b> {
         } = self.curr_loc()
         {
             self.set_instrument_type_at(mode, self.curr_loc());
+        } else {
+            panic!("Should have gotten component location!")
+        }
+    }
+
+    fn set_instrument_type_at(&mut self, mode: InstrumentationMode, loc: Location) {
+        if let Location::Component {
+            mod_idx,
+            func_idx,
+            instr_idx,
+        } = loc
+        {
+            match self.comp.modules[mod_idx]
+                .functions
+                .get_mut(func_idx as FunctionID)
+                .kind
+            {
+                FuncKind::Import(_) => panic!("Can't instrument into an imported function!"),
+                FuncKind::Local(ref mut l) => {
+                    set_instrument_type_for_local_func_at(mode, l, instr_idx);
+                }
+            }
         } else {
             panic!("Should have gotten component location!")
         }
@@ -330,12 +272,7 @@ impl<'a, 'b> Iterator<'b> for ComponentIterator<'a, 'b> {
     }
 
     fn before_at(&mut self, loc: Location) -> &mut Self {
-        if let Location::Component {
-            mod_idx: _mod_idx,
-            func_idx: _func_idx,
-            instr_idx: _instr_idx,
-        } = loc
-        {
+        if let Location::Component { .. } = loc {
             self.set_instrument_type_at(InstrumentationMode::Before, loc);
             self
         } else {
@@ -344,12 +281,7 @@ impl<'a, 'b> Iterator<'b> for ComponentIterator<'a, 'b> {
     }
 
     fn after_at(&mut self, loc: Location) -> &mut Self {
-        if let Location::Component {
-            mod_idx: _mod_idx,
-            func_idx: _func_idx,
-            instr_idx: _instr_idx,
-        } = loc
-        {
+        if let Location::Component { .. } = loc {
             self.set_instrument_type_at(InstrumentationMode::After, loc);
             self
         } else {
@@ -371,36 +303,24 @@ impl<'a, 'b> Iterator<'b> for ComponentIterator<'a, 'b> {
         }
     }
 
-    fn set_instrument_type_at(&mut self, mode: InstrumentationMode, loc: Location) {
+    /// Gets the injected instruction at the current location by index
+    fn get_injected_val(&self, idx: usize) -> &Operator {
         if let Location::Component {
             mod_idx,
             func_idx,
             instr_idx,
-        } = loc
+        } = self.comp_iterator.curr_loc()
         {
-            match self.comp.modules[mod_idx]
+            match &self.comp.modules[mod_idx]
                 .functions
-                .get_mut(func_idx as FunctionID)
+                .get(func_idx as FunctionID)
                 .kind
             {
-                FuncKind::Import(_) => panic!("Can't instrument into an imported function!"),
-                FuncKind::Local(ref mut l) => {
-                    if l.body.instructions[instr_idx].1.get_curr()
-                        == InstrumentType::NotInstrumented
-                    {
-                        l.body.instructions[instr_idx].1 = Instrument::Instrumented {
-                            before: vec![],
-                            after: vec![],
-                            alternate: vec![],
-                            current: mode,
-                        }
-                    } else {
-                        l.body.instructions[instr_idx].1.set_curr(mode);
-                    }
-                }
+                FuncKind::Import(_) => panic!("Can't inject into an imported function!"),
+                FuncKind::Local(l) => l.body.instructions[instr_idx].1.get_instr(idx),
             }
         } else {
-            panic!("Should have gotten component location!")
+            panic!("Should have gotten Component Location and not Module Location!")
         }
     }
 
@@ -410,13 +330,58 @@ impl<'a, 'b> Iterator<'b> for ComponentIterator<'a, 'b> {
     }
 }
 
+// Note: Marked Trait as the same lifetime as component
+impl<'a, 'b> Iterator<'b> for ComponentIterator<'a, 'b> {
+
+    /// Resets the Component Iterator
+    fn reset(&mut self) {
+        self.comp_iterator.reset();
+    }
+
+    /// Goes to the next instruction
+    fn next(&mut self) -> Option<&Operator> {
+        match self.comp_iterator.next() {
+            false => None,
+            true => self.curr_op(),
+        }
+    }
+
+    /// Returns the current location
+    fn curr_loc(&self) -> Location {
+        self.comp_iterator.curr_loc()
+    }
+
+    /// Returns the instruction at the current location
+    fn curr_op(&self) -> Option<&Operator> {
+        if self.comp_iterator.end() {
+            None
+        } else if let Location::Component {
+            mod_idx,
+            func_idx,
+            instr_idx,
+        } = self.comp_iterator.curr_loc()
+        {
+            match &self.comp.modules[mod_idx]
+                .functions
+                .get(func_idx as FunctionID)
+                .kind
+            {
+                FuncKind::Import(_) => panic!("Can't inject into an imported function!"),
+                FuncKind::Local(l) => Some(&l.body.instructions[instr_idx].0),
+            }
+        } else {
+            panic!("Should have gotten Component Location and not Module Location!")
+        }
+    }
+}
+
 impl ModuleBuilder for ComponentIterator<'_, '_> {
     fn add_local(&mut self, val_type: DataType) -> LocalID {
         let curr_loc = self.curr_loc();
         if let Location::Component {
             mod_idx,
             func_idx,
-            instr_idx: _,
+            ..
         } = curr_loc
         {
             {
