+++ src/iterator/module_iterator.rs
@@ -5,7 +5,7 @@ use crate::ir::module::module_functions::FuncKind;
 use crate::ir::module::module_globals::Global;
 use crate::ir::module::Module;
 use crate::ir::types::{DataType, Instrument, InstrumentType, InstrumentationMode, Location};
-use crate::iterator::iterator_trait::Iterator;
+use crate::iterator::iterator_trait::{Instrumenter, Iterator, set_instrument_type_for_local_func_at};
 use crate::opcode::{Inject, MacroOpcode, Opcode};
 use crate::subiterator::module_subiterator::ModuleSubIterator;
 use std::collections::HashMap;
@@ -82,7 +82,7 @@ impl<'a, 'b> Inject<'b> for ModuleIterator<'a, 'b> {
     /// use orca::iterator::module_iterator::ModuleIterator;
     /// use wasmparser::Operator;
     /// use orca::ir::types::{Location};
-    /// use orca::iterator::iterator_trait::Iterator;
+    /// use orca::iterator::iterator_trait::{Instrumenter, Iterator};
     /// use orca::opcode::Opcode;
     ///
     /// let file = "path_to_file";
@@ -132,45 +132,7 @@ impl<'a, 'b> Inject<'b> for ModuleIterator<'a, 'b> {
 }
 impl<'a, 'b> Opcode<'b> for ModuleIterator<'a, 'b> {}
 impl<'a, 'b> MacroOpcode<'b> for ModuleIterator<'a, 'b> {}
-
-// Note: Marked Trait as the same lifetime as component
-impl<'a, 'b> Iterator<'b> for ModuleIterator<'a, 'b> {
-    /// Marks the current location as InstrumentBefore
-    fn before(&mut self) -> &mut Self {
-        self.set_instrument_type(InstrumentationMode::Before);
-        self
-    }
-
-    /// Marks the current location as InstrumentAfter
-    fn after(&mut self) -> &mut Self {
-        self.set_instrument_type(InstrumentationMode::After);
-        self
-    }
-
-    /// Marks the current location as InstrumentAlternate
-    fn alternate(&mut self) -> &mut Self {
-        self.set_instrument_type(InstrumentationMode::Alternate);
-        self
-    }
-
-    /// Resets the Module Iterator
-    fn reset(&mut self) {
-        self.mod_iterator.reset();
-    }
-
-    /// Goes to the next instruction and returns the instruction
-    fn next(&mut self) -> Option<&Operator> {
-        match self.mod_iterator.next() {
-            false => None,
-            true => self.curr_op(),
-        }
-    }
-
-    /// Returns the current Location
-    fn curr_loc(&self) -> Location {
-        self.mod_iterator.curr_loc()
-    }
-
+impl <'a, 'b> Instrumenter<'b> for ModuleIterator<'a, 'b> {
     /// Returns the Instrumentation at the current Location
     fn curr_instrument_type(&self) -> InstrumentType {
         if let Location::Module {
@@ -187,47 +149,27 @@ impl<'a, 'b> Iterator<'b> for ModuleIterator<'a, 'b> {
         }
     }
 
-    /// Returns the current instruction
-    fn curr_op(&self) -> Option<&Operator<'b>> {
-        if self.mod_iterator.end() {
-            None
-        } else if let Location::Module {
-            func_idx,
-            instr_idx,
-        } = self.mod_iterator.curr_loc()
-        {
-            match &self.module.functions.get(func_idx as FunctionID).kind {
-                FuncKind::Import(_) => panic!("Cannot get an instruction to an imported function"),
-                FuncKind::Local(l) => Some(&l.body.instructions[instr_idx].0),
-            }
-        } else {
-            panic!("Should have gotten Module Location!")
-        }
-    }
-
-    /// Gets the injected instruction at the current location by index
-    fn get_injected_val(&self, idx: usize) -> &Operator {
-        if let Location::Module {
-            func_idx,
-            instr_idx,
-        } = self.mod_iterator.curr_loc()
-        {
-            match &self.module.functions.get(func_idx as FunctionID).kind {
-                FuncKind::Import(_) => panic!("Cannot get an instruction to an imported function"),
-                FuncKind::Local(l) => l.body.instructions[instr_idx].1.get_instr(idx),
-            }
-        } else {
-            panic!("Should have gotten Component Location and not Module Location!")
-        }
-    }
-
     fn set_instrument_type(&mut self, mode: InstrumentationMode) {
-        if let Location::Module {
-            func_idx: _func_idx,
-            instr_idx: _instr_idx,
-        } = self.curr_loc()
+        if let Location::Module { .. } = self.curr_loc()
         {
             self.set_instrument_type_at(mode, self.curr_loc());
+        } else {
+            panic!("Should have gotten module location!")
+        }
+    }
+
+    fn set_instrument_type_at(&mut self, mode: InstrumentationMode, loc: Location) {
+        if let Location::Module {
+            func_idx,
+            instr_idx,
+        } = loc
+        {
+            match self.module.functions.get_mut(func_idx as FunctionID).kind {
+                FuncKind::Import(_) => panic!("Cannot add an instruction to an imported function"),
+                FuncKind::Local(ref mut l) => {
+                    set_instrument_type_for_local_func_at(mode, l, instr_idx);
+                }
+            }
         } else {
             panic!("Should have gotten module location!")
         }
@@ -263,10 +205,7 @@ impl<'a, 'b> Iterator<'b> for ModuleIterator<'a, 'b> {
     }
 
     fn before_at(&mut self, loc: Location) -> &mut Self {
-        if let Location::Module {
-            func_idx: _func_idx,
-            instr_idx: _instr_idx,
-        } = loc
+        if let Location::Module { .. } = loc
         {
             self.set_instrument_type_at(InstrumentationMode::Before, loc);
             self
@@ -276,10 +215,7 @@ impl<'a, 'b> Iterator<'b> for ModuleIterator<'a, 'b> {
     }
 
     fn after_at(&mut self, loc: Location) -> &mut Self {
-        if let Location::Module {
-            func_idx: _func_idx,
-            instr_idx: _instr_idx,
-        } = loc
+        if let Location::Module { .. } = loc
         {
             self.set_instrument_type_at(InstrumentationMode::After, loc);
             self
@@ -289,10 +225,7 @@ impl<'a, 'b> Iterator<'b> for ModuleIterator<'a, 'b> {
     }
 
     fn alternate_at(&mut self, loc: Location) -> &mut Self {
-        if let Location::Module {
-            func_idx: _func_idx,
-            instr_idx: _instr_idx,
-        } = loc
+        if let Location::Module { .. } = loc
         {
             self.set_instrument_type_at(InstrumentationMode::Alternate, loc);
             self
@@ -301,35 +234,63 @@ impl<'a, 'b> Iterator<'b> for ModuleIterator<'a, 'b> {
         }
     }
 
-    fn set_instrument_type_at(&mut self, mode: InstrumentationMode, loc: Location) {
+    /// Gets the injected instruction at the current location by index
+    fn get_injected_val(&self, idx: usize) -> &Operator {
         if let Location::Module {
             func_idx,
             instr_idx,
-        } = loc
+        } = self.mod_iterator.curr_loc()
         {
-            match self.module.functions.get_mut(func_idx as FunctionID).kind {
-                FuncKind::Import(_) => panic!("Cannot add an instruction to an imported function"),
-                FuncKind::Local(ref mut l) => {
-                    if l.body.instructions[instr_idx].1.get_curr()
-                        == InstrumentType::NotInstrumented
-                    {
-                        l.body.instructions[instr_idx].1 = Instrument::Instrumented {
-                            before: vec![],
-                            after: vec![],
-                            alternate: vec![],
-                            current: mode,
-                        }
-                    } else {
-                        l.body.instructions[instr_idx].1.set_curr(mode);
-                    }
-                }
+            match &self.module.functions.get(func_idx as FunctionID).kind {
+                FuncKind::Import(_) => panic!("Cannot get an instruction to an imported function"),
+                FuncKind::Local(l) => l.body.instructions[instr_idx].1.get_instr(idx),
             }
         } else {
-            panic!("Should have gotten module location!")
+            panic!("Should have gotten Component Location and not Module Location!")
         }
     }
 
     fn add_global(&mut self, global: Global) -> GlobalID {
         self.module.globals.add(global)
+    }
+}
+
+// Note: Marked Trait as the same lifetime as component
+impl<'a, 'b> Iterator<'b> for ModuleIterator<'a, 'b> {
+
+    /// Resets the Module Iterator
+    fn reset(&mut self) {
+        self.mod_iterator.reset();
+    }
+
+    /// Goes to the next instruction and returns the instruction
+    fn next(&mut self) -> Option<&Operator> {
+        match self.mod_iterator.next() {
+            false => None,
+            true => self.curr_op(),
+        }
+    }
+
+    /// Returns the current Location
+    fn curr_loc(&self) -> Location {
+        self.mod_iterator.curr_loc()
+    }
+
+    /// Returns the current instruction
+    fn curr_op(&self) -> Option<&Operator<'b>> {
+        if self.mod_iterator.end() {
+            None
+        } else if let Location::Module {
+            func_idx,
+            instr_idx,
+        } = self.mod_iterator.curr_loc()
+        {
+            match &self.module.functions.get(func_idx as FunctionID).kind {
+                FuncKind::Import(_) => panic!("Cannot get an instruction to an imported function"),
+                FuncKind::Local(l) => Some(&l.body.instructions[instr_idx].0),
+            }
+        } else {
+            panic!("Should have gotten Module Location!")
+        }
     }
 }
